fix matrix alignment issue during create_calldata
When creating the calldata, we didn't calculate the size of the matrix
correctly because there won't be alignment issue in some platforms, so
we will need to add paddings to those cases. Similarly, we will need to
removing the paddings when reading back from underlying buffer.

And this has to be done in create_calldata and read_calldata in
valueref.py becuase those call will finally writting or reading the
raw device buffer where has no layout info. Therefore, we have to
organize the data layout correctly before writing and after reading.