Fix race condition in hot reload test
Sometimes after re-writing the shader, it did return "2" instead of "1" in the run_and_verify call right after it, so the shader was already reloaded. This is because on command buffer submission, we check for changed shaders and recreate sessions if a shader has changed (enabled because m_auto_detect_changes is true). This is a race condition, because under normal (slow) circumstances, FileSystemWatcher's thread doesn't execute between write_shader and run_and_verify calls, but in rare cases it does and thus triggers the reload.

The race condition can be reproduced by adding a sleep after the write_shader call.

The fix is to not check that the shader hasn't reloaded yet as that would only work with m_auto_detect_changes == false, instead we just check that it is actually reloaded.